+++ dbt_platform_helper/commands/environment.py
@@ -7,8 +7,8 @@ from dbt_platform_helper.domain.copilot_environment import CopilotEnvironment
 from dbt_platform_helper.domain.maintenance_page import MaintenancePageProvider
 from dbt_platform_helper.domain.terraform_environment import TerraformEnvironment
 from dbt_platform_helper.platform_exception import PlatformException
+from dbt_platform_helper.providers.config import ConfigProvider
 from dbt_platform_helper.utils.click import ClickDocOptGroup
-from dbt_platform_helper.utils.validation import load_and_validate_platform_config
 from dbt_platform_helper.utils.versioning import (
     check_platform_helper_version_needs_update,
 )
@@ -65,7 +65,8 @@ def generate(name, vpc_name):
         )
         raise click.Abort
     try:
-        conf = load_and_validate_platform_config()
+        config_provider = ConfigProvider()
+        conf = config_provider.load_and_validate_platform_config()
     except SchemaError as ex:
         click.secho(f"Invalid `{PLATFORM_CONFIG_FILE}` file: {str(ex)}", fg="red")
         raise click.Abort
@@ -82,5 +83,6 @@ def generate(name, vpc_name):
     help=f"Override the default version of terraform-platform-modules. (Default version is '{DEFAULT_TERRAFORM_PLATFORM_MODULES_VERSION}').",
 )
 def generate_terraform(name, terraform_platform_modules_version):
-    conf = load_and_validate_platform_config()
+    config_provider = ConfigProvider()
+    conf = config_provider.load_and_validate_platform_config()
     TerraformEnvironment.generate(conf, name, terraform_platform_modules_version)
