+++ tests/platform_helper/providers/test_config.py
@@ -1,11 +1,12 @@
 from pathlib import Path
+from unittest.mock import patch
 
 import pytest
 import yaml
 
 from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.providers.config import ConfigProvider
-from dbt_platform_helper.utils.validation import load_and_validate_platform_config
+from tests.platform_helper.conftest import FIXTURES_DIR
 
 
 def test_lint_yaml_for_duplicate_keys_fails_when_duplicate_keys_provided(
@@ -44,9 +45,301 @@ extensions:
     assert expected_error in linting_failures[0]
 
     with pytest.raises(SystemExit) as excinfo:
-        load_and_validate_platform_config(PLATFORM_CONFIG_FILE)
+        config_provider.load_and_validate_platform_config(PLATFORM_CONFIG_FILE)
 
     captured = capsys.readouterr()
 
     assert expected_error in captured.err
     assert excinfo.value.code == 1
+
+
+@pytest.mark.parametrize("pipeline_to_trigger", ("", "non-existent-pipeline"))
+@patch("dbt_platform_helper.providers.config.abort_with_error")
+def test_validate_platform_config_fails_if_pipeline_to_trigger_not_valid(
+    mock_abort_with_error, valid_platform_config, pipeline_to_trigger
+):
+    valid_platform_config["environment_pipelines"]["main"][
+        "pipeline_to_trigger"
+    ] = pipeline_to_trigger
+
+    config_provider = ConfigProvider(valid_platform_config)
+
+    config_provider.validate_platform_config()
+    message = mock_abort_with_error.call_args.args[0]
+
+    assert "The following pipelines are misconfigured:" in message
+    assert (
+        f"  'main' - '{pipeline_to_trigger}' is not a valid target pipeline to trigger" in message
+    )
+
+
+@patch("dbt_platform_helper.providers.config.abort_with_error")
+def test_validate_platform_config_fails_with_multiple_errors_if_pipeline_to_trigger_is_invalid(
+    mock_abort_with_error, valid_platform_config
+):
+    valid_platform_config["environment_pipelines"]["main"]["pipeline_to_trigger"] = ""
+    valid_platform_config["environment_pipelines"]["test"][
+        "pipeline_to_trigger"
+    ] = "non-existent-pipeline"
+
+    config_provider = ConfigProvider(valid_platform_config)
+
+    config_provider.validate_platform_config()
+    message = mock_abort_with_error.call_args.args[0]
+
+    assert "The following pipelines are misconfigured:" in message
+    assert f"  'main' - '' is not a valid target pipeline to trigger" in message
+    assert (
+        f"  'test' - 'non-existent-pipeline' is not a valid target pipeline to trigger" in message
+    )
+
+
+@patch("dbt_platform_helper.providers.config.abort_with_error")
+def test_validate_platform_config_fails_if_pipeline_to_trigger_is_triggering_itself(
+    mock_abort_with_error, valid_platform_config
+):
+    valid_platform_config["environment_pipelines"]["main"]["pipeline_to_trigger"] = "main"
+    config_provider = ConfigProvider(valid_platform_config)
+    config_provider.validate_platform_config()
+    message = mock_abort_with_error.call_args.args[0]
+
+    assert "The following pipelines are misconfigured:" in message
+    assert f"  'main' - pipelines cannot trigger themselves" in message
+
+
+@pytest.mark.parametrize(
+    "account, envs, exp_bad_envs",
+    [
+        ("account-does-not-exist", ["dev"], ["dev"]),
+        ("prod-acc", ["dev", "staging", "prod"], ["dev", "staging"]),
+        ("non-prod-acc", ["dev", "prod"], ["prod"]),
+    ],
+)
+@patch("dbt_platform_helper.providers.config.abort_with_error")
+def test_validate_platform_config_fails_if_pipeline_account_does_not_match_environment_accounts_with_single_pipeline(
+    mock_abort_with_error, platform_env_config, account, envs, exp_bad_envs
+):
+    platform_env_config["environment_pipelines"] = {
+        "main": {
+            "account": account,
+            "slack_channel": "/codebuild/notification_channel",
+            "trigger_on_push": True,
+            "environments": {env: {} for env in envs},
+        }
+    }
+
+    config_provider = ConfigProvider(platform_env_config)
+
+    config_provider.validate_platform_config()
+
+    message = mock_abort_with_error.call_args.args[0]
+
+    assert "The following pipelines are misconfigured:" in message
+    assert (
+        f"  'main' - these environments are not in the '{account}' account: {', '.join(exp_bad_envs)}"
+        in message
+    )
+
+
+@patch("dbt_platform_helper.providers.config.abort_with_error")
+def test_validate_platform_config_fails_if_database_copy_config_is_invalid(
+    mock_abort_with_error,
+):
+    """Edge cases for this are all covered in unit tests of
+    validate_database_copy_section elsewhere in this file."""
+    config = {
+        "application": "test-app",
+        "environments": {"dev": {}, "test": {}, "prod": {}},
+        "extensions": {
+            "our-postgres": {
+                "type": "postgres",
+                "version": 7,
+                "database_copy": [{"from": "dev", "to": "dev"}],
+            }
+        },
+    }
+
+    config_provider = ConfigProvider(config)
+
+    config_provider.validate_platform_config()
+
+    message = mock_abort_with_error.call_args.args[0]
+
+    assert (
+        f"database_copy 'to' and 'from' cannot be the same environment in extension 'our-postgres'."
+        in message
+    )
+
+
+@patch("dbt_platform_helper.providers.config.abort_with_error")
+def test_validate_platform_config_catches_database_copy_errors(
+    mock_abort_with_error, platform_env_config
+):
+    platform_env_config["environment_pipelines"] = {
+        "main": {
+            "account": "non-prod",
+            "slack_channel": "/codebuild/notification_channel",
+            "trigger_on_push": True,
+            "environments": {"dev": {}, "prod": {}},
+        },
+        "prod": {
+            "account": "prod",
+            "slack_channel": "/codebuild/notification_channel",
+            "trigger_on_push": True,
+            "environments": {"dev": {}, "staging": {}, "prod": {}},
+        },
+    }
+
+    config_provider = ConfigProvider(platform_env_config)
+
+    config_provider.validate_platform_config()
+
+    message = mock_abort_with_error.call_args.args[0]
+
+    assert "The following pipelines are misconfigured:" in message
+    assert f"  'main' - these environments are not in the 'non-prod' account: dev" in message
+    assert f"  'prod' - these environments are not in the 'prod' account: dev, staging" in message
+
+
+@pytest.mark.parametrize(
+    "account, envs",
+    [
+        ("non-prod-acc", ["dev", "staging"]),
+        ("prod-acc", ["prod"]),
+    ],
+)
+def test_validate_platform_config_succeeds_if_pipeline_account_matches_environment_accounts(
+    platform_env_config, account, envs
+):
+    platform_env_config["environment_pipelines"] = {
+        "main": {
+            "account": account,
+            "slack_channel": "/codebuild/notification_channel",
+            "trigger_on_push": True,
+            "environments": {env: {} for env in envs},
+        }
+    }
+
+    # Should not error if config is sound.
+    config_provider = ConfigProvider(platform_env_config)
+
+    config_provider.validate_platform_config()
+
+
+@pytest.mark.parametrize(
+    "yaml_file",
+    [
+        "pipeline/platform-config.yml",
+        "pipeline/platform-config-with-public-repo.yml",
+        "pipeline/platform-config-for-terraform.yml",
+    ],
+)
+def test_load_and_validate_config_valid_file(yaml_file):
+    """Test that, given the path to a valid yaml file, load_and_validate_config
+    returns the loaded yaml unmodified."""
+
+    config_provider = ConfigProvider()
+
+    path = FIXTURES_DIR / yaml_file
+    validated = config_provider.load_and_validate_platform_config(path=path)
+
+    with open(path, "r") as fd:
+        conf = yaml.safe_load(fd)
+
+    assert validated == conf
+
+
+def test_validation_fails_if_invalid_default_version_keys_present(
+    fakefs, capsys, valid_platform_config
+):
+    valid_platform_config["default_versions"] = {"something-invalid": "1.2.3"}
+    Path(PLATFORM_CONFIG_FILE).write_text(yaml.dump(valid_platform_config))
+
+    config_provider = ConfigProvider(valid_platform_config)
+
+    with pytest.raises(SystemExit) as ex:
+        config_provider.load_and_validate_platform_config()
+
+        assert "Wrong key 'something-invalid'" in str(ex)
+
+
+@pytest.mark.parametrize(
+    "invalid_key",
+    (
+        "",
+        "invalid-key",
+        "platform-helper",  # platform-helper is not valid in the environment overrides.
+    ),
+)
+def test_validation_fails_if_invalid_environment_version_override_keys_present(
+    invalid_key, fakefs, capsys, valid_platform_config
+):
+    valid_platform_config["environments"]["*"]["versions"] = {invalid_key: "1.2.3"}
+    Path(PLATFORM_CONFIG_FILE).write_text(yaml.dump(valid_platform_config))
+    config_provider = ConfigProvider(valid_platform_config)
+
+    with pytest.raises(SystemExit) as ex:
+        config_provider.load_and_validate_platform_config()
+
+        assert f"Wrong key '{invalid_key}'" in str(ex)
+
+
+@pytest.mark.parametrize(
+    "invalid_key",
+    (
+        "",
+        "invalid-key",
+        "terraform-platform-modules",  # terraform-platform-modules is not valid in the pipeline overrides.
+    ),
+)
+def test_validation_fails_if_invalid_pipeline_version_override_keys_present(
+    invalid_key, fakefs, capsys, valid_platform_config
+):
+    valid_platform_config["environment_pipelines"]["test"]["versions"][invalid_key] = "1.2.3"
+    Path(PLATFORM_CONFIG_FILE).write_text(yaml.dump(valid_platform_config))
+    config_provider = ConfigProvider(valid_platform_config)
+
+    with pytest.raises(SystemExit) as ex:
+        config_provider.load_and_validate_platform_config()
+
+        assert f"Wrong key '{invalid_key}'" in str(ex)
+
+
+def test_load_and_validate_platform_config_fails_with_invalid_yaml(fakefs, capsys):
+    """Test that, given the path to an invalid yaml file,
+    load_and_validate_config aborts and prints an error."""
+
+    Path(PLATFORM_CONFIG_FILE).write_text("{invalid data")
+    config_provider = ConfigProvider()
+    with pytest.raises(SystemExit):
+        config_provider.load_and_validate_platform_config()
+
+    assert f"Error: {PLATFORM_CONFIG_FILE} is not valid YAML" in capsys.readouterr().err
+
+
+def test_validation_runs_against_platform_config_yml(fakefs):
+    fakefs.create_file(PLATFORM_CONFIG_FILE, contents='{"application": "my_app"}')
+    config_provider = ConfigProvider()
+    config = config_provider.load_and_validate_platform_config(path=PLATFORM_CONFIG_FILE)
+
+    assert list(config.keys()) == ["application"]
+    assert config["application"] == "my_app"
+
+
+def test_aws_validation_can_be_switched_off(s3_extensions_fixture, capfd):
+    config_provider = ConfigProvider()
+    config_provider.load_and_validate_platform_config()
+
+    assert "Warning" not in capfd.readouterr().out
+
+
+def test_load_and_validate_platform_config_skips_file_check_when_disable_file_check_parameter_passed(
+    capfd, fakefs
+):
+    fakefs.create_file(PLATFORM_CONFIG_FILE, contents=yaml.dump({"application": "my_app"}))
+    config_provider = ConfigProvider()
+
+    with patch.object(ConfigProvider, "config_file_check") as mock_config_file_check:
+        config_provider.load_and_validate_platform_config(disable_file_check=True)
+
+        assert not mock_config_file_check.called
