+++ tests/platform_helper/utils/test_validation.py
@@ -1,19 +1,12 @@
 import re
 from pathlib import Path
-from unittest.mock import MagicMock
-from unittest.mock import patch
 
 import pytest
 import yaml
 from schema import SchemaError
 
-from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
-from dbt_platform_helper.providers.config import PlatformConfigValidator
 from dbt_platform_helper.providers.platform_config_schema import PlatformConfigSchema
-from dbt_platform_helper.utils.validation import load_and_validate_platform_config
 from dbt_platform_helper.utils.validation import validate_addons
-from dbt_platform_helper.utils.validation import validate_platform_config
-from tests.platform_helper.conftest import FIXTURES_DIR
 from tests.platform_helper.conftest import UTILS_FIXTURES_DIR
 
 
@@ -352,625 +345,3 @@ def test_validate_s3_bucket_name_multiple_failures():
     ]
     for exp_error in exp_errors:
         assert exp_error in str(ex.value)
-
-
-@pytest.mark.parametrize("pipeline_to_trigger", ("", "non-existent-pipeline"))
-@patch("dbt_platform_helper.providers.config.abort_with_error")
-def test_validate_platform_config_fails_if_pipeline_to_trigger_not_valid(
-    mock_abort_with_error, valid_platform_config, pipeline_to_trigger
-):
-    valid_platform_config["environment_pipelines"]["main"][
-        "pipeline_to_trigger"
-    ] = pipeline_to_trigger
-
-    validate_platform_config(valid_platform_config)
-    message = mock_abort_with_error.call_args.args[0]
-
-    assert "The following pipelines are misconfigured:" in message
-    assert (
-        f"  'main' - '{pipeline_to_trigger}' is not a valid target pipeline to trigger" in message
-    )
-
-
-@patch("dbt_platform_helper.providers.config.abort_with_error")
-def test_validate_platform_config_fails_with_multiple_errors_if_pipeline_to_trigger_is_invalid(
-    mock_abort_with_error, valid_platform_config
-):
-    valid_platform_config["environment_pipelines"]["main"]["pipeline_to_trigger"] = ""
-    valid_platform_config["environment_pipelines"]["test"][
-        "pipeline_to_trigger"
-    ] = "non-existent-pipeline"
-
-    validate_platform_config(valid_platform_config)
-    message = mock_abort_with_error.call_args.args[0]
-
-    assert "The following pipelines are misconfigured:" in message
-    assert f"  'main' - '' is not a valid target pipeline to trigger" in message
-    assert (
-        f"  'test' - 'non-existent-pipeline' is not a valid target pipeline to trigger" in message
-    )
-
-
-@patch("dbt_platform_helper.providers.config.abort_with_error")
-def test_validate_platform_config_fails_if_pipeline_to_trigger_is_triggering_itself(
-    mock_abort_with_error, valid_platform_config
-):
-    valid_platform_config["environment_pipelines"]["main"]["pipeline_to_trigger"] = "main"
-
-    validate_platform_config(valid_platform_config)
-    message = mock_abort_with_error.call_args.args[0]
-
-    assert "The following pipelines are misconfigured:" in message
-    assert f"  'main' - pipelines cannot trigger themselves" in message
-
-
-@pytest.mark.parametrize(
-    "account, envs, exp_bad_envs",
-    [
-        ("account-does-not-exist", ["dev"], ["dev"]),
-        ("prod-acc", ["dev", "staging", "prod"], ["dev", "staging"]),
-        ("non-prod-acc", ["dev", "prod"], ["prod"]),
-    ],
-)
-@patch("dbt_platform_helper.providers.config.abort_with_error")
-def test_validate_platform_config_fails_if_pipeline_account_does_not_match_environment_accounts_with_single_pipeline(
-    mock_abort_with_error, platform_env_config, account, envs, exp_bad_envs
-):
-    platform_env_config["environment_pipelines"] = {
-        "main": {
-            "account": account,
-            "slack_channel": "/codebuild/notification_channel",
-            "trigger_on_push": True,
-            "environments": {env: {} for env in envs},
-        }
-    }
-
-    validate_platform_config(platform_env_config)
-
-    message = mock_abort_with_error.call_args.args[0]
-
-    assert "The following pipelines are misconfigured:" in message
-    assert (
-        f"  'main' - these environments are not in the '{account}' account: {', '.join(exp_bad_envs)}"
-        in message
-    )
-
-
-@patch("dbt_platform_helper.providers.config.abort_with_error")
-def test_validate_platform_config_fails_if_database_copy_config_is_invalid(
-    mock_abort_with_error,
-):
-    """Edge cases for this are all covered in unit tests of
-    validate_database_copy_section elsewhere in this file."""
-    config = {
-        "application": "test-app",
-        "environments": {"dev": {}, "test": {}, "prod": {}},
-        "extensions": {
-            "our-postgres": {
-                "type": "postgres",
-                "version": 7,
-                "database_copy": [{"from": "dev", "to": "dev"}],
-            }
-        },
-    }
-
-    validate_platform_config(config)
-
-    message = mock_abort_with_error.call_args.args[0]
-
-    assert (
-        f"database_copy 'to' and 'from' cannot be the same environment in extension 'our-postgres'."
-        in message
-    )
-
-
-@patch("dbt_platform_helper.providers.config.abort_with_error")
-def test_validate_platform_config_catches_database_copy_errors(
-    mock_abort_with_error, platform_env_config
-):
-    platform_env_config["environment_pipelines"] = {
-        "main": {
-            "account": "non-prod",
-            "slack_channel": "/codebuild/notification_channel",
-            "trigger_on_push": True,
-            "environments": {"dev": {}, "prod": {}},
-        },
-        "prod": {
-            "account": "prod",
-            "slack_channel": "/codebuild/notification_channel",
-            "trigger_on_push": True,
-            "environments": {"dev": {}, "staging": {}, "prod": {}},
-        },
-    }
-
-    validate_platform_config(platform_env_config)
-
-    message = mock_abort_with_error.call_args.args[0]
-
-    assert "The following pipelines are misconfigured:" in message
-    assert f"  'main' - these environments are not in the 'non-prod' account: dev" in message
-    assert f"  'prod' - these environments are not in the 'prod' account: dev, staging" in message
-
-
-@pytest.mark.parametrize(
-    "account, envs",
-    [
-        ("non-prod-acc", ["dev", "staging"]),
-        ("prod-acc", ["prod"]),
-    ],
-)
-def test_validate_platform_config_succeeds_if_pipeline_account_matches_environment_accounts(
-    platform_env_config, account, envs
-):
-    platform_env_config["environment_pipelines"] = {
-        "main": {
-            "account": account,
-            "slack_channel": "/codebuild/notification_channel",
-            "trigger_on_push": True,
-            "environments": {env: {} for env in envs},
-        }
-    }
-
-    # Should not error if config is sound.
-    validate_platform_config(platform_env_config)
-
-
-@pytest.mark.parametrize(
-    "yaml_file",
-    [
-        "pipeline/platform-config.yml",
-        "pipeline/platform-config-with-public-repo.yml",
-        "pipeline/platform-config-for-terraform.yml",
-    ],
-)
-def test_load_and_validate_config_valid_file(yaml_file):
-    """Test that, given the path to a valid yaml file, load_and_validate_config
-    returns the loaded yaml unmodified."""
-
-    path = FIXTURES_DIR / yaml_file
-    validated = load_and_validate_platform_config(path=path)
-
-    with open(path, "r") as fd:
-        conf = yaml.safe_load(fd)
-
-    assert validated == conf
-
-
-def test_validation_fails_if_invalid_default_version_keys_present(
-    fakefs, capsys, valid_platform_config
-):
-    valid_platform_config["default_versions"] = {"something-invalid": "1.2.3"}
-    Path(PLATFORM_CONFIG_FILE).write_text(yaml.dump(valid_platform_config))
-
-    with pytest.raises(SystemExit) as ex:
-        load_and_validate_platform_config()
-
-        assert "Wrong key 'something-invalid'" in str(ex)
-
-
-@pytest.mark.parametrize(
-    "invalid_key",
-    (
-        "",
-        "invalid-key",
-        "platform-helper",  # platform-helper is not valid in the environment overrides.
-    ),
-)
-def test_validation_fails_if_invalid_environment_version_override_keys_present(
-    invalid_key, fakefs, capsys, valid_platform_config
-):
-    valid_platform_config["environments"]["*"]["versions"] = {invalid_key: "1.2.3"}
-    Path(PLATFORM_CONFIG_FILE).write_text(yaml.dump(valid_platform_config))
-
-    with pytest.raises(SystemExit) as ex:
-        load_and_validate_platform_config()
-
-        assert f"Wrong key '{invalid_key}'" in str(ex)
-
-
-@pytest.mark.parametrize(
-    "invalid_key",
-    (
-        "",
-        "invalid-key",
-        "terraform-platform-modules",  # terraform-platform-modules is not valid in the pipeline overrides.
-    ),
-)
-def test_validation_fails_if_invalid_pipeline_version_override_keys_present(
-    invalid_key, fakefs, capsys, valid_platform_config
-):
-    valid_platform_config["environment_pipelines"]["test"]["versions"][invalid_key] = "1.2.3"
-    Path(PLATFORM_CONFIG_FILE).write_text(yaml.dump(valid_platform_config))
-
-    with pytest.raises(SystemExit) as ex:
-        load_and_validate_platform_config()
-
-        assert f"Wrong key '{invalid_key}'" in str(ex)
-
-
-def test_load_and_validate_platform_config_fails_with_invalid_yaml(fakefs, capsys):
-    """Test that, given the path to an invalid yaml file,
-    load_and_validate_config aborts and prints an error."""
-
-    Path(PLATFORM_CONFIG_FILE).write_text("{invalid data")
-    with pytest.raises(SystemExit):
-        load_and_validate_platform_config()
-
-    assert f"Error: {PLATFORM_CONFIG_FILE} is not valid YAML" in capsys.readouterr().err
-
-
-def test_validation_runs_against_platform_config_yml(fakefs):
-    fakefs.create_file(PLATFORM_CONFIG_FILE, contents='{"application": "my_app"}')
-
-    config = load_and_validate_platform_config()
-
-    assert list(config.keys()) == ["application"]
-    assert config["application"] == "my_app"
-
-
-def test_aws_validation_can_be_switched_off(s3_extensions_fixture, capfd):
-    load_and_validate_platform_config()
-
-    assert "Warning" not in capfd.readouterr().out
-
-
-@patch("dbt_platform_helper.utils.validation.config_file_check")
-def test_load_and_validate_platform_config_skips_file_check_when_disable_file_check_parameter_passed(
-    mock_config_file_check, capfd, fakefs
-):
-    fakefs.create_file(PLATFORM_CONFIG_FILE, contents=yaml.dump({"application": "my_app"}))
-    load_and_validate_platform_config(disable_file_check=True)
-
-    assert not mock_config_file_check.called
-
-
-@pytest.mark.parametrize(
-    "database_copy_section",
-    [
-        None,
-        [{"from": "dev", "to": "test"}],
-        [{"from": "test", "to": "dev"}],
-        [
-            {
-                "from": "prod",
-                "to": "test",
-                "from_account": "9999999999",
-                "to_account": "1122334455",
-            }
-        ],
-        [
-            {
-                "from": "dev",
-                "to": "test",
-                "from_account": "9999999999",
-                "to_account": "9999999999",
-            }
-        ],
-        [{"from": "test", "to": "dev", "pipeline": {}}],
-        [{"from": "test", "to": "dev", "pipeline": {"schedule": "0 0 * * WED"}}],
-        [
-            {
-                "from": "test",
-                "to": "dev",
-                "from_account": "9999999999",
-                "to_account": "1122334455",
-                "pipeline": {"schedule": "0 0 * * WED"},
-            }
-        ],
-    ],
-)
-def test_validate_database_copy_section_success_cases(database_copy_section):
-    config = {
-        "application": "test-app",
-        "environments": {
-            "dev": {"accounts": {"deploy": {"id": "1122334455"}}},
-            "test": {"accounts": {"deploy": {"id": "1122334455"}}},
-            "prod": {"accounts": {"deploy": {"id": "9999999999"}}},
-        },
-        "extensions": {
-            "our-postgres": {
-                "type": "postgres",
-                "version": 7,
-            }
-        },
-    }
-
-    if database_copy_section:
-        config["extensions"]["our-postgres"]["database_copy"] = database_copy_section
-
-    PlatformConfigValidator.validate_database_copy_section(config)
-
-    # Should get here fine if the config is valid.
-
-
-@pytest.mark.parametrize(
-    "database_copy_section, expected_parameters",
-    [
-        ([{"from": "hotfix", "to": "test"}], ["from"]),
-        ([{"from": "dev", "to": "hotfix"}], ["to"]),
-        ([{"from": "hotfix", "to": "hotfix"}], ["to", "from"]),
-        ([{"from": "test", "to": "dev"}, {"from": "dev", "to": "hotfix"}], ["to"]),
-        ([{"from": "hotfix", "to": "test"}, {"from": "dev", "to": "test"}], ["from"]),
-    ],
-)
-def test_validate_database_copy_section_failure_cases(
-    capfd, database_copy_section, expected_parameters
-):
-    config = {
-        "application": "test-app",
-        "environments": {"dev": {}, "test": {}, "prod": {}},
-        "extensions": {
-            "our-postgres": {
-                "type": "postgres",
-                "version": 7,
-            }
-        },
-    }
-
-    config["extensions"]["our-postgres"]["database_copy"] = database_copy_section
-
-    with pytest.raises(SystemExit):
-        PlatformConfigValidator.validate_database_copy_section(config)
-
-    console_message = capfd.readouterr().err
-
-    for param in expected_parameters:
-        msg = f"database_copy '{param}' parameter must be a valid environment (dev, test, prod) but was 'hotfix' in extension 'our-postgres'."
-        assert msg in console_message
-
-
-def test_validate_database_copy_fails_if_from_and_to_are_the_same(capfd):
-    config = {
-        "application": "test-app",
-        "environments": {"dev": {}, "test": {}, "prod": {}},
-        "extensions": {
-            "our-postgres": {
-                "type": "postgres",
-                "version": 7,
-                "database_copy": [{"from": "dev", "to": "dev"}],
-            }
-        },
-    }
-
-    with pytest.raises(SystemExit):
-        PlatformConfigValidator.validate_database_copy_section(config)
-
-    console_message = capfd.readouterr().err
-
-    msg = (
-        f"database_copy 'to' and 'from' cannot be the same environment in extension 'our-postgres'."
-    )
-    assert msg in console_message
-
-
-@pytest.mark.parametrize(
-    "env_name",
-    ["prod", "prod-env", "env-that-is-prod", "thing-prod-thing"],
-)
-def test_validate_database_copy_section_fails_if_the_to_environment_is_prod(capfd, env_name):
-    config = {
-        "application": "test-app",
-        "environments": {"dev": {}, "test": {}, "prod": {}},
-        "extensions": {
-            "our-postgres": {
-                "type": "postgres",
-                "version": 7,
-                "database_copy": [{"from": "dev", "to": env_name}],
-            }
-        },
-    }
-
-    with pytest.raises(SystemExit):
-        PlatformConfigValidator.validate_database_copy_section(config)
-
-    console_message = capfd.readouterr().err
-
-    msg = f"Copying to a prod environment is not supported: database_copy 'to' cannot be '{env_name}' in extension 'our-postgres'."
-    assert msg in console_message
-
-
-def test_validate_database_copy_multi_postgres_success():
-    config = {
-        "application": "test-app",
-        "environments": {"dev": {}, "test": {}, "prod": {}},
-        "extensions": {
-            "our-postgres": {
-                "type": "postgres",
-                "version": 7,
-                "database_copy": [{"from": "dev", "to": "test"}],
-            },
-            "our-other-postgres": {
-                "type": "postgres",
-                "version": 7,
-                "database_copy": [{"from": "dev", "to": "test"}, {"from": "prod", "to": "dev"}],
-            },
-        },
-    }
-
-    PlatformConfigValidator.validate_database_copy_section(config)
-
-    # Should get here fine if the config is valid.
-
-
-def test_validate_database_copy_multi_postgres_failures(capfd):
-    config = {
-        "application": "test-app",
-        "environments": {"dev": {}, "test": {}, "prod": {}},
-        "extensions": {
-            "our-postgres": {
-                "type": "postgres",
-                "version": 7,
-                "database_copy": [{"from": "devvv", "to": "test"}],
-            },
-            "our-other-postgres": {
-                "type": "postgres",
-                "version": 7,
-                "database_copy": [{"from": "test", "to": "test"}, {"from": "dev", "to": "prod"}],
-            },
-        },
-    }
-
-    with pytest.raises(SystemExit):
-        PlatformConfigValidator.validate_database_copy_section(config)
-
-    console_message = capfd.readouterr().err
-
-    assert (
-        f"database_copy 'from' parameter must be a valid environment (dev, test, prod) but was 'devvv' in extension 'our-postgres'."
-        in console_message
-    )
-    assert (
-        f"database_copy 'to' and 'from' cannot be the same environment in extension 'our-other-postgres'."
-        in console_message
-    )
-    assert (
-        f"Copying to a prod environment is not supported: database_copy 'to' cannot be 'prod' in extension 'our-other-postgres'."
-        in console_message
-    )
-
-
-def test_validate_database_copy_fails_if_cross_account_with_no_from_account(capfd):
-    config = {
-        "application": "test-app",
-        "environments": {
-            "dev": {"accounts": {"deploy": {"id": "1122334455"}}},
-            "prod": {"accounts": {"deploy": {"id": "9999999999"}}},
-        },
-        "extensions": {
-            "our-postgres": {
-                "type": "postgres",
-                "version": 7,
-                "database_copy": [{"from": "prod", "to": "dev"}],
-            }
-        },
-    }
-
-    with pytest.raises(SystemExit):
-        PlatformConfigValidator.validate_database_copy_section(config)
-
-    console_message = capfd.readouterr().err
-
-    msg = f"Environments 'prod' and 'dev' are in different AWS accounts. The 'from_account' parameter must be present."
-    assert msg in console_message
-
-
-def test_validate_database_copy_fails_if_cross_account_with_no_to_account(capfd):
-    config = {
-        "application": "test-app",
-        "environments": {
-            "dev": {"accounts": {"deploy": {"id": "1122334455"}}},
-            "prod": {"accounts": {"deploy": {"id": "9999999999"}}},
-        },
-        "extensions": {
-            "our-postgres": {
-                "type": "postgres",
-                "version": 7,
-                "database_copy": [{"from": "prod", "to": "dev", "from_account": "9999999999"}],
-            }
-        },
-    }
-
-    with pytest.raises(SystemExit):
-        PlatformConfigValidator.validate_database_copy_section(config)
-
-    console_message = capfd.readouterr().err
-
-    msg = f"Environments 'prod' and 'dev' are in different AWS accounts. The 'to_account' parameter must be present."
-    assert msg in console_message
-
-
-def test_validate_database_copy_fails_if_cross_account_with_incorrect_account_ids(capfd):
-    config = {
-        "application": "test-app",
-        "environments": {
-            "dev": {"accounts": {"deploy": {"id": "1122334455"}}},
-            "prod": {"accounts": {"deploy": {"id": "9999999999"}}},
-        },
-        "extensions": {
-            "our-postgres": {
-                "type": "postgres",
-                "version": 7,
-                "database_copy": [
-                    {
-                        "from": "prod",
-                        "to": "dev",
-                        "from_account": "000000000",
-                        "to_account": "1111111111",
-                    }
-                ],
-            }
-        },
-    }
-
-    with pytest.raises(SystemExit):
-        PlatformConfigValidator.validate_database_copy_section(config)
-
-    console_message = capfd.readouterr().err
-
-    msg = f"Incorrect value for 'from_account' for environment 'prod'"
-    assert msg in console_message
-
-
-@pytest.mark.parametrize(
-    "config, expected_response",
-    [
-        (
-            # No engine defined in either env
-            {
-                "extensions": {
-                    "connors-redis": {
-                        "type": "redis",
-                        "environments": {"*": {"plan": "tiny"}, "prod": {"plan": "largish"}},
-                    }
-                },
-            },
-            "",
-        ),
-        (
-            # Valid engine version defined in *
-            {
-                "extensions": {
-                    "connors-redis": {
-                        "type": "redis",
-                        "environments": {
-                            "*": {"engine": "7.1", "plan": "tiny"},
-                            "prod": {"plan": "tiny"},
-                        },
-                    }
-                },
-            },
-            "",
-        ),
-        (
-            # Invalid engine defined in prod environment
-            {
-                "extensions": {
-                    "connors-redis": {
-                        "type": "redis",
-                        "environments": {
-                            "*": {"plan": "tiny"},
-                            "prod": {"engine": "invalid", "plan": "tiny"},
-                        },
-                    }
-                },
-            },
-            "redis version for environment prod is not in the list of supported redis versions: ['7.1']. Provided Version: invalid",
-        ),
-    ],
-)
-def test_validate_extension_supported_versions(config, expected_response, capsys):
-
-    mock_redis_provider = MagicMock()
-    mock_redis_provider.get_supported_redis_versions.return_value = ["7.1"]
-
-    PlatformConfigValidator.validate_extension_supported_versions(
-        config=config,
-        extension_type="redis",
-        version_key="engine",
-        get_supported_versions=mock_redis_provider.get_supported_redis_versions,
-    )
-
-    captured = capsys.readouterr()
-
-    assert expected_response in captured.out
-    assert captured.err == ""
